add async/await and copmletion network method
